requirements.txt
PyYAML==5.4.1
termcolor==1.1.0
